requirements.txt
Flask==3.1.1
Flask-PyMongo==2.3.0
Flask-Login==0.6.3
Flask-WTF==1.2.2
WTForms==3.1.2
pymongo==4.14.0
bcrypt==4.3.0
python-dotenv==1.0.0
requests==2.32.4
Pillow
